requirements.txt
flask
sqlalchemy
pytaglib
flask-socketio